sql_api/app/requirements.txt
colorama==0.4.6
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dotenv==1.0.0
